+++ src/chiapos/post.h
@@ -46,13 +46,13 @@ int GetBaseIters(int nTargetHeight, Consensus::Params const& params);
 
 double GetDifficultyChangeMaxFactor(int nTargetHeight, Consensus::Params const& params);
 
-bool AddLocalVdfRequest(uint256 const& challenge, int nIters);
+bool AddLocalVdfRequest(uint256 const& challenge, uint64_t nIters);
 
-std::set<int> QueryLocalVdfRequests(uint256 const& challenge);
+std::set<uint64_t> QueryLocalVdfRequests(uint256 const& challenge);
 
 bool AddLocalVdfProof(CVdfProof vdfProof);
 
-bool FindLocalVdfProof(uint256 const& challenge, int nIters, CVdfProof* pvdfProof = nullptr);
+bool FindLocalVdfProof(uint256 const& challenge, uint64_t nIters, CVdfProof* pvdfProof = nullptr);
 
 std::vector<CVdfProof> QueryLocalVdfProof(uint256 const& challenge);
